Add MYSQL docker image (#4)
This PR adds MYSQL database to docker-image and put some changes:

the docker-compose.yml build the .jar and run springboot itself. It is no more needed to built it ourselves and put it into the container.
the docker-compose.yml mount the database in the mysql image with the provided .sql file
NOTE: If you change the .sql file, you have to store it in the mysql-dump folder.
No more Dockerfile needed
run the entire project:

docker-compose up

IMPORTANT NOTE: You have to run the project with docker-compose and not anymore with your IDE or in command line. Otherwise, the database will fail to init.
OR, if you don't want to use docker-compose for any reason, you have to change the values in application.properties file BUT DON'T COMMIT those changes.

Also. this PR provides a working .sql script